pyproject.toml
[project]
name = "mysql-db-backups"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-api-python-client==2.108.0",
    "google-auth==2.25.2",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "python-dotenv==1.0.0",
]
